Decompose search_sauna.py for abstraction
