+++ src/smart_sauna_map/scraper/abstract_scraper.py
@@ -1,0 +1,6 @@
+from abc import ABC
+
+
+class AbstractScraper(ABC):
+    # TODO: とりあえず箱だけ作ったのでメソッド・プロパティを埋めていく
+    pass
