+++ src/smart_sauna_map/search_sauna.py
@@ -1,55 +1,13 @@
 # -*- coding: utf-8 -*-
 from __future__ import annotations
 
-import abc
-import re
-from dataclasses import dataclass
 from functools import cache
 from typing import Optional
-from urllib.parse import urlencode, urlparse
 
-import requests
-from bs4 import BeautifulSoup
-
-from smart_sauna_map.geocoding import geocode
+from smart_sauna_map.sauna import Sauna
+from smart_sauna_map.scraper.sauna_ikitai_scraper import SaunaIkitaiScraper
 
 __all__ = ["search_sauna"]
-
-
-@dataclass
-class Sauna:
-    sauna_id: int
-    name: str
-    address: str
-    ikitai: int
-    lat: float | None
-    lng: float | None
-    image_url: str | None
-    mans_room: MansRoom | None
-    womans_room: WomansRoom | None
-    unisex_room: UnisexRoom | None
-    description: list[str] | None
-
-
-@dataclass
-class BathRoom(abc.ABC):
-    sauna_temperature: float | None
-    mizuburo_temperature: float | None
-
-
-@dataclass
-class MansRoom(BathRoom):
-    pass
-
-
-@dataclass
-class WomansRoom(BathRoom):
-    pass
-
-
-@dataclass
-class UnisexRoom(BathRoom):
-    pass
 
 
 @cache
@@ -90,181 +48,12 @@ def search_sauna(
             )
         ]
     """
-    response: str = _request(keyword, prefecture, page=page_index)
-    soup: BeautifulSoup = _parse(response)
-    return _extract_saunas(soup)
-
-
-def _request(
-    keyword: Optional[str] = "富士",
-    prefecture: Optional[str] = "tokyo",
-    page: Optional[int] = 1,
-) -> str:
-    url = "https://sauna-ikitai.com/search"
-    payload: dict[str, str | int] = {}
-    if keyword:
-        payload.update({"keyword": keyword})
-    if prefecture:
-        payload.update({"prefecture[]": prefecture})
-    if page:
-        payload.update({"page": page})
-
-    res: requests.models.Response = _sub_request(url, payload)
-    _raise_error_if_status_code_is_not_200(res)
-    return res.text
-
-
-def _sub_request(
-    url: str, payload: dict[str, str | int], timeout: float = 3.0
-) -> requests.models.Response:
-    return requests.get(url, params=urlencode(payload), timeout=timeout)
-
-
-def _raise_error_if_status_code_is_not_200(res: requests.models.Response):
-    res.raise_for_status()
-
-
-def _parse(res: str) -> BeautifulSoup:
-    return BeautifulSoup(res, "html.parser")
-
-
-def _extract_saunas(soup: BeautifulSoup) -> list[Sauna]:
-    def sauna(s: BeautifulSoup) -> Sauna:
-        name = _extract_sauna_name(s)
-        latlng = geocode(name)
-        return Sauna(
-            sauna_id=_extract_sauna_id(s),
-            name=name,
-            address=_extract_sauna_address(s),
-            ikitai=_extract_sauna_ikitai(s),
-            lat=latlng.get("lat", None),
-            lng=latlng.get("lng", None),
-            image_url=_extract_sauna_image_url(s),
-            mans_room=_extract_sauna_mans_room(s),
-            womans_room=_extract_sauna_womans_room(s),
-            unisex_room=_extract_sauna_unisex_room(s),
-            description=_extract_sauna_description(s),
-        )
-
-    return [sauna(s) for s in soup.find_all(class_="p-saunaItem")]
-
-
-def _extract_sauna_name(soup: BeautifulSoup) -> str:
-    def parse(s):
-        return s.find("h3").text.strip()
-
-    return parse(soup.find(class_="p-saunaItemName"))
-
-
-def _extract_sauna_address(soup: BeautifulSoup) -> str:
-    def parse(s):
-        return s.text.strip().replace("\xa0", "")
-
-    return parse(soup.find("address"))
-
-
-def _extract_sauna_id(soup: BeautifulSoup) -> int:
-    def parse(s):
-        url = s.find(name="a").get("href")
-        sauna_id = int(urlparse(url).path.split("/")[-1])
-        return sauna_id
-
-    return parse(soup)
-
-
-def _extract_sauna_image_url(soup: BeautifulSoup) -> str:
-    def parse(s):
-        url = str(s.find(name="img").get("src"))
-        return url
-
-    return parse(soup.find(class_="p-saunaItem_image"))
-
-
-def _extract_sauna_mans_room(soup: BeautifulSoup) -> MansRoom | None:
-    s = soup.find(class_="p-saunaItemSpec_content p-saunaItemSpec_content--man")
-    if not s:
-        return None
-    return MansRoom(
-        sauna_temperature=_find_sauna_temperature(s),
-        mizuburo_temperature=_find_mizuburo_temperature(s),
+    scraper = SaunaIkitaiScraper()
+    return scraper.search_sauna(
+        keyword=keyword,
+        prefecture=prefecture,
+        page_index=page_index,
     )
-
-
-def _extract_sauna_womans_room(soup: BeautifulSoup) -> WomansRoom | None:
-    s = soup.find(class_="p-saunaItemSpec_content p-saunaItemSpec_content--woman")
-    if not s:
-        return None
-    return WomansRoom(
-        sauna_temperature=_find_sauna_temperature(s),
-        mizuburo_temperature=_find_mizuburo_temperature(s),
-    )
-
-
-def _extract_sauna_unisex_room(soup: BeautifulSoup) -> UnisexRoom | None:
-    s = soup.find(class_="p-saunaItemSpec_content p-saunaItemSpec_content--unisex")
-    if not s:
-        return None
-    return UnisexRoom(
-        sauna_temperature=_find_sauna_temperature(s),
-        mizuburo_temperature=_find_mizuburo_temperature(s),
-    )
-
-
-def _find_sauna_temperature(s):
-    return _extract_digits(
-        s.find(class_="p-saunaItemSpec_item p-saunaItemSpec_item--sauna")
-        .find(class_="p-saunaItemSpec_value")
-        .text
-    )
-
-
-def _find_mizuburo_temperature(s):
-    return _extract_digits(
-        s.find(class_="p-saunaItemSpec_item p-saunaItemSpec_item--mizuburo")
-        .find(class_="p-saunaItemSpec_value")
-        .text
-    )
-
-
-def _extract_digits(text: str) -> float | None:
-    matched = re.match(r"\d+(\.\d+)?", text)
-    if not matched:
-        return None
-    return float(matched.group())
-
-
-def _extract_sauna_description(soup: BeautifulSoup) -> list[str]:
-    def parse(s):
-        descs = []
-        for _s in s.find_all(name="li"):
-            descs.append(_s.text)
-        return descs
-
-    return parse(soup.find(class_="p-saunaItem_informations"))
-
-
-def _extract_sauna_ikitai(
-    soup, class_: str = "p-saunaItem_actions", search_string: str = "イキタイ"
-) -> int:
-    ikitai = None
-    contents = soup.find(class_=class_).contents
-
-    for content in contents:
-        if search_string in content.text:
-            matched_texts = re.findall(r"[\d]+", content.text)
-            if len(matched_texts) > 1:
-                raise ValueError(
-                    f"ikitai number must have a digit, but found multiple"
-                    f" digits {matched_texts}."
-                )
-            ikitai = int(matched_texts[0])
-
-    if ikitai is None:
-        raise ValueError(
-            f"ikitai number must be in content, but not found in {contents}."
-        )
-
-    return ikitai
 
 
 if __name__ == "__main__":
